+++ src/main/java/webserver/Controller.java
@@ -1,0 +1,7 @@
+package webserver;
+
+@FunctionalInterface
+interface Controller {
+    Response execute(HttpRequest httpRequest);
+
+}
