+++ src/main/java/webserver/ControllerNotFoundException.java
@@ -1,0 +1,7 @@
+package webserver;
+
+class ControllerNotFoundException extends RuntimeException {
+    public ControllerNotFoundException(String uri) {
+        super(String.format("Not Found: %s", uri));
+    }
+}
